flyttade till rätt fil och det fungerade woho
